refactor(desktop): extract bot incoming flow
